Add rough Swagger
This is a rough swagger definition that I'm not precious about, just intended as a starting point for (1) making sure we're all agreed on the endpoints and (2) making sure we've covered all the bases. The domain parts are unfinished, but the initial work is focused on the Account aspects.

To render the Swagger YAML, https://editor-next.swagger.io/ is pretty good.

Signed-off-by: Blaine Cook <romeda@gmail.com>