feat: expand /healthcheck to verify that the db schema is up to date
When the server is started, we query the database to find the
version of the latest applied migration. The healthcheck endpoint
then performs this query again, to verify that both versions are
equal: otherwise, it means new migrations have been performed,
and the server needs to be restarted using the newer schema.