chore: update `config` from `0.13` to `0.14` and MSRV to 1.75
We need config 0.14 for nightly tests to successfully compile.
If not, we'd transitively depend on `ahash@0.7.7`, which doesn't compile under latest nightly.

I also updated all minor versions of dependencies, which pushed up
our MSRV due to dependencies up to 1.74, and so I thought whatever and
updated it to 1.75 since that gives us the cool async trait stuff.

Also update the nix flake to rustc version 1.76.